# refactor: dont covert to charcode in iswhitespace