--- a/lib/path.js
+++ b/lib/path.js
@@ -3,6 +3,7 @@ import { removeLeadingZero, toFixed } from './svgo/tools.js';
 /**
  * @typedef {import('./types.js').PathDataItem} PathDataItem
  * @typedef {import('./types.js').PathDataCommand} PathDataCommand
+ * @typedef {'none' | 'sign' | 'whole' | 'decimal_point' | 'decimal' | 'e' | 'exponent_sign' | 'exponent'} ReadNumberState
  */
 
 // Based on https://www.w3.org/TR/SVG11/paths.html#PathDataBNF
@@ -38,20 +39,16 @@ const isCommand = (c) => {
 };
 
 /**
- * @type {(c: string) => boolean}
+ * @param {string} c
+ * @returns {boolean}
  */
-const isWsp = (c) => {
-  const codePoint = c.codePointAt(0);
-  return (
-    codePoint === 0x20 ||
-    codePoint === 0x9 ||
-    codePoint === 0xd ||
-    codePoint === 0xa
-  );
+const isWhiteSpace = (c) => {
+  return c === ' ' || c === '\t' || c === '\r' || c === '\n';
 };
 
 /**
- * @type {(c: string) => boolean}
+ * @param {string} c
+ * @returns {boolean}
  */
 const isDigit = (c) => {
   const codePoint = c.codePointAt(0);
@@ -60,10 +57,6 @@ const isDigit = (c) => {
   }
   return 48 <= codePoint && codePoint <= 57;
 };
-
-/**
- * @typedef {'none' | 'sign' | 'whole' | 'decimal_point' | 'decimal' | 'e' | 'exponent_sign' | 'exponent'} ReadNumberState
- */
 
 /**
  * @type {(string: string, cursor: number) => [number, ?number]}
@@ -133,7 +126,8 @@ const readNumber = (string, cursor) => {
 };
 
 /**
- * @type {(string: string) => PathDataItem[]}
+ * @param {string} string
+ * @returns {PathDataItem[]}
  */
 export const parsePathData = (string) => {
   /**
@@ -150,7 +144,7 @@ export const parsePathData = (string) => {
   let hadComma = false;
   for (let i = 0; i < string.length; i += 1) {
     const c = string.charAt(i);
-    if (isWsp(c)) {
+    if (isWhiteSpace(c)) {
       continue;
     }
     // allow comma only between arguments
@@ -170,11 +164,9 @@ export const parsePathData = (string) => {
         if (c !== 'M' && c !== 'm') {
           return pathData;
         }
-      } else {
+      } else if (args.length !== 0) {
         // stop if previous command arguments are not flushed
-        if (args.length !== 0) {
-          return pathData;
-        }
+        return pathData;
       }
       command = c;
       args = [];